fix(useAnimations): flag to enable sleep wake scenarios
